로그인 인증 및, 회원 기능 추가 예외사항 알림 추가

+++ front/src/pages/Detail.jsx
@@ -3,13 +3,13 @@ import { Navigate, useNavigate, useParams } from "react-router-dom";
 import axios from "axios";
 
 import { Cookies } from "react-cookie";
-import Swal from "sweetalert2";
-import withReactContent from "sweetalert2-react-content";
-import { setPost } from "../store/store";
+import { setLogIn, setPost } from "../store/store";
 import base_url from "../config/BaseUrl";
 import { deleteHomePostAPI } from "../api/post";
 
-const MySwal = withReactContent(Swal);
+// sweetalert2 : alerts 디자인 라이브러리
+import Swal from "sweetalert2";
+import { checkLogInAPI } from "../api/user";
 
 axios.defaults.baseURL = base_url;
 
@@ -27,14 +27,54 @@ function Detail(props) {
     }
   }
 
+  const loginWarning = () => {
+    Swal.fire({
+      icon: "warning",
+      title: "로그인이 필요합니다.",
+    });
+  };
+
+  const authorWarning = () => {
+    Swal.fire({
+      icon: "warning",
+      title: "본인이 작성한 글만 삭제 가능합니다.",
+    });
+  };
+
+  const deleteWarning = () => {
+    Swal.fire({
+      title: "정말 삭제하시겠습니까?",
+      text: "한 번 삭제하면 다시 되돌릴 수 없습니다.",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "삭제!",
+      cancelButtonText: "취소",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire("삭제!", "글이 성공적으로 삭제 되었습니다.", "success");
+        deleteHomePostAPI(id)
+          .then(() => {
+            navigate("/");
+          })
+          .catch((err) => {
+            console.log(err);
+          });
+      }
+    });
+  };
   const onClick = () => {
-    deleteHomePostAPI(id)
-      .then(() => {
-        navigate("/");
-      })
-      .catch((err) => {
-        console.log(err);
-      });
+    checkLogInAPI().then((res) => {
+      // console.log(res.data);
+      if (!res.data.logIn) {
+        loginWarning();
+      } else if (res.data.id != finded_post.id) {
+        authorWarning();
+      } else {
+        deleteWarning();
+      }
+    });
   };
 
   return (
